move additional reports files
